Better typing through project restructuring
